Board Space Class
Added Scrabble file to contain 'ScrabbleNamespace', created class boardSpace in file boardSpaceClass, renamed tile file to tileClass - tile class name remains the same
