+++ Assets/Classes/Scrabble.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 71c03bd08bff4cf45b87945b81f80ec5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
