+++ Assets/Classes/boardSpaceClass.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3eccf2f2aef5edf4cbc9fa364f633f3f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
